refactor(app): batch7 — extract useGlobalShortcutHandlers
